Fixed references. Added code lens support.
References were not working before because the last 2 capabilities sent to the client is 'prepareRename' and 'references'. Since prepareRename isn't an actual capability, it would fail before reaching references. The fix was to simply remove IPrepareRename from RenameHandler, which apparently wan't being used anyway.
Clicking on the 'references' link in the code lens does not work yet.