Fixed virtual lookup table for a certain scenario.
class A {
  public virtual void func() {}
}
class B : A {
  public override void func() {}
}
class C : B {
}
new C().func();

In the code above, A.func() is called rather than C.func(). Since C does not override func(), it would resolve to the default, which is A.func() when it should be B.func(). This fix will cause B.func() to run instead.