Return types, parameter types, and variable types are now an ICodeTypeSolver rather than an ICodeType.
With the explicit changes, workshop functions can return the Player type or have parameters that require the Player type. The Player type requires the player variable scope which is only initialized when parsing the user script, so we need to initialize the workshop IMethods every time we parse, costing around 100 milliseconds. ICodeTypeSolver will instead grab the CodeType from the DeltinScript so we don't need to initialize the IMethods every time. CodeType implements ICodeTypeSolver which will just solve to itself.

Since this change makes GetLabel require DeltinScript, I decided to clean up the GetLabel code while I was at it. Which was overdue, anyway.