Bakemap node index is encoded into the vector itself
Getting the index of the closest node to a point is an important part of pathfinding. The previous version looked like this (workshop pseudocode):
```
node_index = Nodes.IndexOf(Nodes.Sort(node => node.DistanceTo(point)))
```
This isn't particularly good. The `Nodes` array is searched through twice, once for the closest node and again for the index of that node.

This commit solves this by encoding the index of the node into the vector itself. The index is encoded into the 4th, 5th, and 6th decimal places of the Y component. For example, a pathmap looking like this:
```
[{10.6, 34.2, 5.9}, {1.2, 6.2423, 90.34}, {7.5, 45.3, 2.3}, ...]
```
will be encoded as:
```
[{10.6, 34.2, 5.9}, {1.2, 6.242001, 90.34}, {7.5, 45.300002, 2.3}, ...]
```
So we only need to do this without the extraneous index search:
```
node = Nodes.Sort(node => node.DistanceTo(point))
```
Then to extract the index, we do this:
```
// turn the 4th, 5th, and 6th decimal place into a single whole number
node_index = (node.Y * 100 - floor(node.Y * 100)) * 100
```

Additionally, this commit converts the Vertex class into a struct, which is more suitable. This has side effects for the currently unused Model stuff.