Fixed continue/break in repeated function calls
If you had a function definition like this:
```cs
void my_func() {
    for(Number i; 10; 1) {
        if (i == 5) {
            break;
        }
    }
}
```
And called it multiple times:
```cs
my_func();
my_func();
```
The function would be written to the workshop output twice.
However, the first continue would skip to the end of the *next* loop in the second function call.
This bug occured because continues and breaks were tracked by the semantics checker rather than the action set instance.