Lambdas capture local macros.
This was done for 2 reasons.
The first is to provide a way to capture a global variable without wasting a workshop variable just to invoke ostw's variable capturing capabilities.
The second is that local macros can use elements that do not exist in the lambda's context, so this also fixes some bugs during compilation.

This commit fixes a couple of other bugs that were highlighted by this change:
- Fixed macros in classes incorrectly creating class object stacks.
- Fixed macros in classes not working in classes with type parameters.